requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.5
numpy==1.24.3
scikit-learn==1.3.2
aiofiles==23.2.1
streamlit==1.28.1

